requirements.txt
streamlit
openai
python-dotenv
ffmpeg-python
